requirements_app.txt
# HyperCopa DISEC 2026 — Dependências do app entregável
#
# Versões pinadas para garantir REPRODUTIBILIDADE IDÊNTICA da entrega à banca.
# Testado com Python 3.10, 3.11, 3.12.
#
# H2O e scikit-learn estão pinados em versão exata porque mudanças entre
# minor versions podem alterar leaderboard, métricas e variável-importância.
#
# Java JDK 17 ou 21 é REQUISITO EXTERNO do H2O (não vem no pip).
# Verifique antes de instalar:    java -version

# UI e dados
streamlit~=1.39.0
pandas~=2.2.3
numpy~=1.26.4
plotly~=5.24.1
pyarrow~=17.0.0
polars~=1.12.0

# Modelagem (PIN EXATO — afeta resultado)
h2o==3.46.0.6
scikit-learn==1.5.2

# Agente IA (Caminho A)
openai~=1.54.0
python-dotenv~=1.0.1

# Geração de documentos (.docx)
python-docx~=1.1.2

# API REST opcional (Caminho B / integração futura)
fastapi~=0.115.4
uvicorn~=0.32.0
pydantic~=2.9.2
